remove unused import and rename validate to validateJWT
