+++ lib/index.js
@@ -1,8 +1,6 @@
 'use strict';
-const JWT       = require('jsonwebtoken');
 const unless    = require('koa-unless');
-
-const verify    = require('./verify');
+const verifyJWT = require('./verify-jwt');
 
 module.exports = opts => {
   opts = opts || {};
@@ -32,7 +30,7 @@ module.exports = opts => {
       ctx.throw(401, 'Invalid secret\n');
     }
 
-    return verify(token, secret, opts)
+    return verifyJWT(token, secret, opts)
       .then(isRevoked(ctx, token))
       .then(user => {
         ctx.state = ctx.state || {};
