+++ lib/verify-jwt.js
